colGroup {reactable}	R Documentation

Column group definitions

Description

Use colGroup() to create column groups in a table.

Usage

colGroup(
  name = NULL,
  columns = NULL,
  header = NULL,
  html = FALSE,
  align = NULL,
  headerVAlign = NULL,
  sticky = NULL,
  headerClass = NULL,
  headerStyle = NULL
)

Arguments

name
Column group header name.

columns
Character vector of column names in the group.

header
Custom header renderer. An R function that takes the header value as an argument, or a JS() function that takes a column object and table state object as arguments.

html
Render header content as HTML? Raw HTML strings are escaped by default.

align
Horizontal alignment of content in the column group header. One of "left", "right", "center" (the default).

headerVAlign
Vertical alignment of content in the column group header. One of "top" (the default), "center", "bottom".

sticky
Make the column group sticky when scrolling horizontally? Either "left" or "right" to make the column group stick to the left or right side.

If a column group is sticky, all columns in the group will automatically be made sticky.

headerClass
Additional CSS classes to apply to the header.

headerStyle
Inline styles to apply to the header. A named list or character string.

Note that if headerStyle is a named list, property names should be camelCased.

Value

A column group definition object that can be used to create column groups in reactable().

Examples

reactable(
  iris,
  columns = list(
    Sepal.Length = colDef(name = "Length"),
    Sepal.Width = colDef(name = "Width"),
    Petal.Length = colDef(name = "Length"),
    Petal.Width = colDef(name = "Width")
  ),
  columnGroups = list(
    colGroup(name = "Sepal", columns = c("Sepal.Length", "Sepal.Width")),
    colGroup(name = "Petal", columns = c("Petal.Length", "Petal.Width"))
  )
)

[Package reactable version 0.4.5 Index]