reactable-package {reactable}	R Documentation

reactable: Interactive Data Tables for R

Description

Interactive data tables for R, based on the 'React Table' JavaScript library. Provides an HTML widget that can be used in 'R Markdown' or 'Quarto' documents, 'Shiny' applications, or viewed from an R console.

Details

See the online documentation for examples and an extensive usage guide.

Author(s)

Maintainer: Greg Lin glin@glin.io

Other contributors:

Tanner Linsley (React Table library) [contributor, copyright holder]

Emotion team and other contributors (Emotion library) [contributor, copyright holder]

Kent Russell (reactR package) [contributor, copyright holder]

Ramnath Vaidyanathan (htmlwidgets package) [contributor, copyright holder]

Joe Cheng (htmlwidgets package) [contributor, copyright holder]

JJ Allaire (htmlwidgets package) [contributor, copyright holder]

Yihui Xie (htmlwidgets package) [contributor, copyright holder]

Kenton Russell (htmlwidgets package) [contributor, copyright holder]

Facebook, Inc. and its affiliates (React library) [contributor, copyright holder]

FormatJS (FormatJS libraries) [contributor, copyright holder]

Feross Aboukhadijeh, and other contributors (buffer library) [contributor, copyright holder]

Roman Shtylman (process library) [contributor, copyright holder]

James Halliday (stream-browserify library) [contributor, copyright holder]

Posit Software, PBC [funder, copyright holder]

See Also

Useful links:

https://glin.github.io/reactable/

https://github.com/glin/reactable

Report bugs at https://github.com/glin/reactable/issues

[Package reactable version 0.4.5 Index]