Column group definitions

Column group definitions

colGroup(name = NULL, columns = NULL, header = NULL, html = FALSE,
  align = NULL, headerClass = NULL, headerStyle = NULL)

Arguments

name
Column group header name.

columns
Character vector of column names in the group.

header
Custom header renderer. An R function that takes the header value as an argument, or a JS() function that takes a column info object as an argument.

html
Render header as HTML? HTML strings are escaped by default.

align
Column group header alignment. One of "left", "right", "center".

headerClass
Additional CSS classes to apply to the header.

headerStyle
Inline styles to apply to the header. A named list or character string.

Note that if headerStyle is a named list, property names should be camelCased.

Value

A column group definition object that can be used to create column groups in reactable().

Examples

reactable(
  iris,
  columns = list(
    Sepal.Length = colDef(name = "Length"),
    Sepal.Width = colDef(name = "Width"),
    Petal.Length = colDef(name = "Length"),
    Petal.Width = colDef(name = "Width")
  ),
  columnGroups = list(
    colGroup(name = "Sepal", columns = c("Sepal.Length", "Sepal.Width")),
    colGroup(name = "Petal", columns = c("Petal.Length", "Petal.Width"))
  )
)