reactable: Interactive Data Tables Based on 'React Table'

Interactive data tables for R, based on the 'React Table' JavaScript library. Provides an HTML widget that can be used in 'R Markdown' documents and 'Shiny' applications, or viewed from an R console.

Details

See the online documentation for examples and an extensive usage guide.

See also

Author

Maintainer: Greg Lin glin@glin.io

Other contributors:

Tanner Linsley (React Table library) [contributor, copyright holder]

Emotion team and other contributors (Emotion library) [contributor, copyright holder]